+++ extensions-compose/api/jvm/extensions-compose.api
@@ -1,11 +1,3 @@
-public final class com/arkivanov/decompose/extensions/compose/PredictiveBackGestureIconKt {
-	public static final fun PredictiveBackGestureIcon-eaDK9VM (Landroidx/compose/ui/graphics/vector/ImageVector;FJJLandroidx/compose/runtime/Composer;II)V
-}
-
-public final class com/arkivanov/decompose/extensions/compose/PredictiveBackGestureOverlayKt {
-	public static final fun PredictiveBackGestureOverlay (Lcom/arkivanov/essenty/backhandler/BackDispatcher;Lkotlin/jvm/functions/Function4;Landroidx/compose/ui/Modifier;ZZLkotlin/jvm/functions/Function0;Lkotlin/jvm/functions/Function2;Landroidx/compose/runtime/Composer;II)V
-}
-
 public final class com/arkivanov/decompose/extensions/compose/SubscribeAsStateKt {
 	public static final fun subscribeAsState (Lcom/arkivanov/decompose/value/Value;Landroidx/compose/runtime/SnapshotMutationPolicy;Landroidx/compose/runtime/Composer;II)Landroidx/compose/runtime/State;
 }
@@ -142,5 +134,13 @@ public final class com/arkivanov/decompose/extensions/compose/stack/animation/pr
 public final class com/arkivanov/decompose/extensions/compose/stack/animation/predictiveback/PredictiveBackAnimationKt {
 	public static final fun predictiveBackAnimation (Lcom/arkivanov/essenty/backhandler/BackHandler;Lcom/arkivanov/decompose/extensions/compose/stack/animation/StackAnimation;Lkotlin/jvm/functions/Function3;Lkotlin/jvm/functions/Function0;)Lcom/arkivanov/decompose/extensions/compose/stack/animation/StackAnimation;
 	public static synthetic fun predictiveBackAnimation$default (Lcom/arkivanov/essenty/backhandler/BackHandler;Lcom/arkivanov/decompose/extensions/compose/stack/animation/StackAnimation;Lkotlin/jvm/functions/Function3;Lkotlin/jvm/functions/Function0;ILjava/lang/Object;)Lcom/arkivanov/decompose/extensions/compose/stack/animation/StackAnimation;
+}
+
+public final class com/arkivanov/decompose/extensions/compose/stack/animation/predictiveback/PredictiveBackGestureIconKt {
+	public static final fun PredictiveBackGestureIcon-eaDK9VM (Landroidx/compose/ui/graphics/vector/ImageVector;FJJLandroidx/compose/runtime/Composer;II)V
+}
+
+public final class com/arkivanov/decompose/extensions/compose/stack/animation/predictiveback/PredictiveBackGestureOverlayKt {
+	public static final fun PredictiveBackGestureOverlay (Lcom/arkivanov/essenty/backhandler/BackDispatcher;Lkotlin/jvm/functions/Function4;Landroidx/compose/ui/Modifier;ZZLkotlin/jvm/functions/Function0;Lkotlin/jvm/functions/Function2;Landroidx/compose/runtime/Composer;II)V
 }
