+++ sample/shared/compose/src/iosMain/kotlin/com/arkivanov/sample/shared/root/RootViewController.kt
@@ -6,8 +6,8 @@ import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.window.ComposeUIViewController
 import com.arkivanov.decompose.ExperimentalDecomposeApi
-import com.arkivanov.decompose.extensions.compose.PredictiveBackGestureIcon
-import com.arkivanov.decompose.extensions.compose.PredictiveBackGestureOverlay
+import com.arkivanov.decompose.extensions.compose.stack.animation.predictiveback.PredictiveBackGestureIcon
+import com.arkivanov.decompose.extensions.compose.stack.animation.predictiveback.PredictiveBackGestureOverlay
 import com.arkivanov.essenty.backhandler.BackDispatcher
 import platform.UIKit.UIViewController
